services/auth_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
bcrypt<4.0.0
